requirements.txt
pyTelegramBotAPI==4.15.4
psycopg2-binary
requests
Pillow
arabic-reshaper
python-bidi
python-dotenv
